Initial commit on splitting out get_sg_eos to a separate compilation unit. This is the first step to moving that code out of sg-eos.
